refactor option handling
